Refactored context, added some extra info
